pyproject.toml
[project]
name = "seek-case-study"
version = "0.0.0"
requires-python = ">=3.12,<3.13"
dependencies = [
    "pandas ~= 2.3.3",
    "streamlit ~= 1.50.0",
    "plotly ~= 6.3.1",
    "pinecone ~= 7.0.1",
    "packaging ~=25.0",
    "strands-agents~=1.8.0",
    "openai",
    "PyPDF2",
    "python-docx"


]

[project.optional-dependencies]
dev = [
    "notebook"
]